daraz-project/requirements.txt
requests
beautifulsoup4
pandas
numpy
scikit-learn
umap-learn
flask
gunicorn
python-dotenv
joblib
PyYAML
